Big hackfest going on for USB ISO transfer.
Seems that we can optimize this a bit but I have hacked on it for a while
and seem to only have made a tiny bit of progress. My goal is to increase
the transfer size to max out the USB bus so we can try to get bigger camera
frames.