Memory leaks on deleteDevice and createDevice
Freenect<>::deleteDevice called map::erase but the device
object pointer was not deleted.

Similarly Freenect<>::createDevice did not consider
the case when you already have the id in the map.
In this case the object is also leaked because the pointer
is just overwritten.

It is likely that the second leak has more implication
regarding the device allocation but the patch, at least
deals with object deallocation.

Signed-off-by: David García Garzón <david.garcia@barcelonamedia.org> (vokimon)