1. Updated record to output the mode struct used.  This is informative, will be changed in the future to save as metadata, and is helpful for setting the default struct in Fakenect. 2. Added fakenect functions that correspond to the new ones in the API.  Note that fakenect only works for the previous default resolution/mode at the moment.
Future Change (as a heads up to other devs)
A later update will store the modes in metadata with the dump and just ensure that if you are setting things that they belong to the recorded mode.  The current behavior will be the default for people with datasets made before this.
Signed-off-by: Brandyn A. White <bwhite@dappervision.com>