Make freenect_frame_mode have well-defined field sizes.
This commit gives explicit sizes to the elements of freenect_frame_mode,
forces the enums contained in freenect_frame_mode to be assigned at least
a 32-bit integer, and for memory packing efficiency, orders the contents
from largest to smallest.

It also adjusts the Python bindings and fakenect/record to take these changes
into account.

Signed-off-by: Drew Fisher <drew.m.fisher@gmail.com>