Add (partial) audio support.
This commit supports streaming of audio from the four microphones.
We still need to implement timestamps, user-provided audio (for
cancellation), and firmware upload support.  But if you have already
loaded the firmware (say, with NUIse), this works great.

Due to the way we're handling isochronous transfers in libusbemu, this
will probably not work on Windows yet.

To enable audio, add -DBUILD_AUDIO to your cmake command line or enable
the BUILD_AUDIO option in the cmake gui.

Signed-off-by: Drew Fisher <drew.m.fisher@gmail.com>