Add support for fetching the audio firmware.
This commit adds a python script to download and extract the audio firmware.
Since we do not have permission to redistribute the audio firmware itself,
we add a build option (BUILD_REDIST_PACKAGE) to disable adding the firmware
to the install and instead shipping the firmware fetcher program.

Packagers should enable this option and add a postinstallation hook that
downloads and places the firmware in /usr/share/libfreenect/.

Users should leave this option disabled, but should NOT redistribute packages
or firmware.

Signed-off-by: Drew Fisher <drew.m.fisher@gmail.com>