Add depth to RGB registration via FREENECT_DEPTH_REGISTERED and raw-to-mm via FREENECT_DEPTH_MM
Also add a simple demo showing overlaid depth and RGB video is included in regview.c

Some advanced users may want to write GPU shaders to implement the transforms,
or want to not apply the transform to the entire frame to save CPU cycles.  For them,
we have added in freenect-registration.h a way to access the internal parameters and shift tables
needed to perform the transform.  Since this includes heap variables, such users should free
the tables when no longer in use with freenect_destroy_registration().

Signed-off-by: Drew Fisher <drew.m.fisher@gmail.com>
Signed-off-by: Kyle McDonald <kyle@kylemcdonald.net>
Signed-off-by: Florian Echtler <floe@butterbrot.org>