Attempt to rework audio/firmware upload.
It turns out being way uglier and longer, but should behave slightly
more correctly in the presence of multiple Kinects.

That said, since there's no reliable way to tell for sure which motor,
camera, and audio device belong to the same physical package, it's still
a little bit of a guessing game.

Signed-off-by: Drew Fisher <drew.m.fisher@gmail.com>