Remove libusb.h include from libfreenect.h
typedef freenect_usb_context as void, as it is already for win32. That way,
we drop the need to have a properly installed developpement files of libusb
when we don't use it directly. This fix also partly the problem of
discrepancy between the include paths of libusb on freebsd and libusb-1.0

Signed-off-by: Nicolas Bourdaud <nicolas.bourdaud@gmail.com>