Fix cmake libusb search and fix libusb include
Adjust cmake test for libusb library so that the compatible libusb for
freebsd can be found.  Since freebsd's libusb header are located in a
different place from libusb-1.0, the includes in the source code must be
adjusted using #include <libusb.h>.  Actually those new include statements
are those that libusb-1.0 really expects to be used (see libusb-1.0.pc
cflags to verify this). With this commit, libfreenect is buildable on
freebsd.

Signed-off-by: Nicolas Bourdaud <nicolas.bourdaud@gmail.com>