Making C++ wrapper compatible with C++11.
Compiling with a simple sample app with C++11 fails. This is due to the changes
to the std::make_pair function. More details are available here:

http://www.cplusplus.com/reference/utility/make_pair/
http://stackoverflow.com/questions/9641960/c11-make-pair-with-specified-template-parameters-doesnt-compile

The applied change means that the code compiles correctly for both C++03 and
C++11. make_pair was never necessary, since the types are explicitly specified
anyway.